requirements.txt
dashscope==1.13.3
python-dotenv==1.0.0 